+++ src/main/java/inventarios/com/Sistema/Inventarios/Repositories/UserInventoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface UserRepository extends JpaRepository<UserInventory, Long> {
+public interface UserInventoryRepository extends JpaRepository<UserInventory, Long> {
     UserInventory findByLogin(String login);
 }
