+++ src/main/java/inventarios/com/Sistema/Inventarios/Services/UserInventoryService.java
@@ -1,10 +1,10 @@
-package inventarios.com.Sistema.Inventarios.Service;
+package inventarios.com.Sistema.Inventarios.Services;
 
 import inventarios.com.Sistema.Inventarios.Models.UserInventory;
 
 import java.util.List;
 
-public interface UserService {
+public interface UserInventoryService {
     void inputUser(UserInventory userInventory);
 
     void modifyUser(UserInventory userInventory);
