+++ src/main/resources/application.properties
@@ -8,9 +8,9 @@ spring.data.rest.base-path=/rest
 
 spring.jackson.serialization.INDENT_OUTPUT=true
 
-#spring.datasource.url=jdbc:postgresql://localhost:5432/homebanking
+#spring.datasource.url=jdbc:postgresql://localhost:5432/inventorySystem
 #
-#spring.datasource.username=homebankingapp
+#spring.datasource.username=inventorySystem
 #
 #spring.datasource.password=12345678
 #
